Add ESLint, initial TypeScript support
This is a huge one for me :) It's just so happened that HBN was my only project where I've already started using JS frameworks confidently, but still didn't know TypeScript at all. Currently it's super raw but at least is compileable, I plan to polish it in a few next days.